Added team win cases
Added 4 different win cases based on surviving players and capture points and a scene displaying the case at the end. I destroy the DontDestroyOnLoad object cause we don't need it anymore, but PlayerMaster still tries to access it (healthBar.GetComponent<HealthBarScript1>().SetHealth(cp.GetHealth());) so it causes an error. This will be fixed in the next push hopefully